Makefile, restructering directories.

+++ inc/bsp/adc.h
@@ -14,7 +14,7 @@ uint8_t AdcIsBusy(void);
 void AdcPinEnable(uint8_t port, uint8_t portNumber);
 void AdcPinDisable(uint8_t port, uint8_t portNumber);
 inline uint16_t AdcGetResult(void);
-uint8_t AdcSample(uint8_t ch);
+uint16_t AdcSample(uint8_t ch);
 void AdcChSelect(AdcChannel_t channel);
